+++ services/streammanager/kinesis/kinesis_suite_test.go
@@ -13,9 +13,9 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/rudderlabs/rudder-go-kit/logger/mock_logger"
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
 	mock_kinesis "github.com/rudderlabs/rudder-server/mocks/services/streammanager/kinesis"
-	mock_logger "github.com/rudderlabs/rudder-server/mocks/utils/logger"
 
 	"github.com/stretchr/testify/assert"