+++ warehouse/internal/api/http_test.go
@@ -13,7 +13,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rudderlabs/rudder-go-kit/stats/memstats"
+	"github.com/rudderlabs/rudder-go-kit/stats"
+
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
 
 	"github.com/stretchr/testify/require"
@@ -178,7 +179,7 @@ func TestAPI_Process(t *testing.T) {
 			wAPI := api.WarehouseAPI{
 				Repo:        r,
 				Logger:      logger.NOP,
-				Stats:       memstats.New(), // TODO: use a NOP stats
+				Stats:       stats.NOP,
 				Multitenant: m,
 			}